Updating tutorial step 11, removing outdated link
It would seem as though "The Groundhog Series of Screencasts" has left the Intenet. Following the link gets a `410 Gone` response, and there's no information about it on the repoze.org site.

Further searching on DuckDuckGo leads to no credible replacement.

Is there any? If not, I would suggest removing that sentence at a minimum. Anything further would be well beyond my realm of expertise, as I'm starting to learn about it. :D